Maybe fix race condition on outbound connection (??)
I noticed that sometimes go-ricochet would make an outbound connection and
authenticate, and then the peer would try to open an im.ricochet.chat
before the Application has had a chance to register its handler for im.ricochet.chat,
which results in the channel request being denied and the Process loop breaking
once extra, leaving the client non-functioning.

It's very hard to reproduce (I think it only happens after the other client
had some tor connection problem? not sure), so I haven't been able to test it,
but it seems like breaking the Process loop immediately instead of in another
goroutine would fix it.

This commit is a strong candidate for reversion in the event of other outbound
connection-related race conditions :)